Implement VersionChange
Ref #25

This implements `VersionChange`, which can take a value of `MajorUpgrade`, `MinorUpgrade`, `Patch`, `Prerelease`, or `Same`. This is calculated automatically comparing the current and previous version and taking in the account of the version scheme.

Using this information, `versionPolicyMimaCheck` will perform either no check, backward check, or both backward and forward check.